Move into properties folder and other code review feedback
